feat: create unified ProjectCard component
